Fix bomb issues and move explosion logic to explosion prefab
fix #8 fix #9
