+++ Assets/Spelunky/Items/Bomb/Bomb.prefab
@@ -125,18 +125,11 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 94de364589954e31b4a8dd7383e7e570, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  explosion: {fileID: 6910937128500511305, guid: 5fde648c9a1f42b4f90fb2bba6094d65,
+  explosion: {fileID: 987747026005618997, guid: 5fde648c9a1f42b4f90fb2bba6094d65,
     type: 3}
   bounceClip: {fileID: 8300000, guid: 39b30b1d23679b24d8015c8d76a31145, type: 3}
   bombTimerClip: {fileID: 8300000, guid: 264515f24014bee45a552a9b68e94d95, type: 3}
-  bombExplosionClip: {fileID: 8300000, guid: 841fe78a659db254fb17c1f7c74e6c4c, type: 3}
   timeToExplode: 3
-  explotionRadius: 16
-  layerMask:
-    serializedVersion: 2
-    m_Bits: 256
-  maxJumpHeight: 24
-  timeToJumpApex: 0.3
 --- !u!82 &-7030764533147636328
 AudioSource:
   m_ObjectHideFlags: 0
